refactor: moved course helper functions into courses/util.ts
