+++ common/courses/notifications.ts
@@ -5,8 +5,7 @@ import { userForPupil } from '../user';
 import * as Prisma from '@prisma/client';
 import { getFirstLecture } from './lectures';
 import { parseSubjectString } from '../util/subjectsutils';
-import { getCourseCapacity, getCourseFreePlaces } from './participants';
-import { getCourseImageURL } from './util';
+import { getCourseCapacity, getCourseFreePlaces, getCourseImageURL } from './util';
 import { getCourse } from '../../graphql/util';
 import { shuffleArray } from '../../common/util/basic';
 import { NotificationContext } from '../notification/types';
